requirements.txt
# Core dependencies
Django
djangorestframework==3.15.0

# Database support
psycopg2-binary==2.9.9  # For PostgreSQL
# mysqlclient==2.2.1    # Uncomment if using MySQL
# sqlite3               # Built into Python; no need to install separately

# Environment management
python-decouple==3.8  # For environment variables

# Caching and async task queue
#django-redis==5.3.0   # For Redis-based caching
#celery==5.3.4         # For async task handling
#redis==5.0.0          # Required by Celery/Redis

# Static files and media
Pillow==10.0.0  # For image handling

# Testing
pytest==7.4.0
pytest-django==4.5.2
factory-boy==3.2.1  # For test data generation

# Security
django-environ==0.11.0  # Alternate environment variable management
#django-axes==5.40.0     # For login attempt monitoring

# Optional but useful packages
django-cors-headers==4.0.0  # For handling CORS in APIs
drf-yasg          # For API documentation
django-debug-toolbar==4.1.0 # For debugging during development

# Version pinning tools
#pip-tools==6.14.0  # Useful for managing dependencies (optional)

# Development-only tools
#ipython==8.17.1  # Enhanced Python shell for debugging
#black==23.9.1    # Code formatter
#flake8==6.2.1    # Linter for Python code

# Deployment
gunicorn  # WSGI HTTP server
whitenoise # Simplify static file serving in production
